add some fixes after review
