+++ libs/app-shell/src/lib/api/urls/url-overrider.pipe.ts
@@ -15,14 +15,19 @@ import { isAbsoluteUrl, resolveUrl } from './url-utils';
        avatarImgSrc: string;
        avatarImgSize = '30x30';
 
+       constructor(
+            @Optional()
+            @Inject(REMOTE_BASE_URL)
+            public baseUrl: string
+       ) {}
+
        ngOnInit(): void {
            this.avatarImgSrc = `assets/avatar-${this.avatarImgSize}.jpg`;
        }
     }
  */
 @Pipe({
-    name: 'fdsUrlOverrider',
-    pure: false
+    name: 'fdsUrlOverrider'
 })
 export class UrlOverriderPipe implements PipeTransform {
     public transform(url: string, baseUrl: string): string {
